# swagger-assets/swagger-spec.json
{"openapi":"3.0.0","info":{"title":"Peer Recognition Software","description":"Peer Recognition API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"Login":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"employee":{"$ref":"#/components/schemas/Users"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["email","password","employee"]},"TagStats":{"type":"object","properties":{"tagstatId":{"type":"number"},"employee":{"$ref":"#/components/schemas/Users"},"tag":{"$ref":"#/components/schemas/Tag"},"countReceived":{"type":"number"},"countSent":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["tagstatId","employee","tag","countReceived","countSent"]},"Report":{"type":"object","properties":{"reportID":{"type":"number"},"employeeFrom":{"$ref":"#/components/schemas/Users"},"recognition":{"$ref":"#/components/schemas/Recognition"},"ReportDate":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["reportID","employeeFrom","recognition","ReportDate"]},"Rockstar":{"type":"object","properties":{"rockstarID":{"type":"number"},"month":{"type":"number"},"year":{"type":"number"},"compID":{"type":"number"},"stats":{"type":"array","items":{"$ref":"#/components/schemas/TagStats"}},"rockstar":{"$ref":"#/components/schemas/Users"},"recognitions":{"type":"array","items":{"$ref":"#/components/schemas/Recognition"}}},"required":["rockstarID","month","year","compID","stats","rockstar","recognitions"]},"Users":{"type":"object","properties":{"company":{"$ref":"#/components/schemas/Company"},"companyId":{"type":"number"},"employeeId":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"positionTitle":{"type":"string"},"isManager":{"type":"boolean"},"role":{"enum":["employee","admin"],"type":"string"},"startDate":{"format":"date-time","type":"string"},"manager":{"$ref":"#/components/schemas/Users"},"managed":{"type":"array","items":{"$ref":"#/components/schemas/Users"}},"recsSent":{"type":"array","items":{"$ref":"#/components/schemas/Recognition"}},"recsReceived":{"type":"array","items":{"$ref":"#/components/schemas/Recognition"}},"numRecsReceived":{"type":"number"},"numRecsSent":{"type":"number"},"login":{"$ref":"#/components/schemas/Login"},"tagStats":{"type":"array","items":{"$ref":"#/components/schemas/TagStats"}},"recsDeleted":{"type":"array","items":{"$ref":"#/components/schemas/Recognition"}},"report":{"type":"array","items":{"$ref":"#/components/schemas/Report"}},"rockstar":{"type":"array","items":{"$ref":"#/components/schemas/Rockstar"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["company","companyId","employeeId","firstName","lastName","positionTitle","isManager","role","startDate","manager","managed","recsSent","recsReceived","numRecsReceived","numRecsSent","login","tagStats","report","rockstar"]},"Comment":{"type":"object","properties":{"commentID":{"type":"number"},"employeeFrom":{"$ref":"#/components/schemas/Users"},"recognition":{"$ref":"#/components/schemas/Recognition"},"CommentDate":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["commentID","employeeFrom","recognition","CommentDate"]},"Reaction":{"type":"object","properties":{"reactionID":{"type":"number"},"employeeFrom":{"$ref":"#/components/schemas/Users"},"recognition":{"$ref":"#/components/schemas/Recognition"},"reactDate":{"format":"date-time","type":"string"},"reactType":{"enum":["like","thumbsUp"],"type":"string"}},"required":["reactionID","employeeFrom","recognition","reactDate","reactType"]},"Recognition":{"type":"object","properties":{"recId":{"type":"number"},"company":{"$ref":"#/components/schemas/Company"},"postDate":{"format":"date-time","type":"string"},"empFrom":{"$ref":"#/components/schemas/Users"},"empTo":{"$ref":"#/components/schemas/Users"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}},"msg":{"type":"string"},"rockstar":{"$ref":"#/components/schemas/Rockstar"},"reports":{"type":"array","items":{"$ref":"#/components/schemas/Report"}},"comments":{"type":"array","items":{"$ref":"#/components/schemas/Comment"}},"reactions":{"type":"array","items":{"$ref":"#/components/schemas/Reaction"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"},"deletedBy":{"$ref":"#/components/schemas/Users"}},"required":["recId","company","postDate","empFrom","empTo","tags","msg","rockstar","reports","comments","reactions"]},"Tag":{"type":"object","properties":{"tagId":{"type":"number"},"value":{"type":"string"},"company":{"$ref":"#/components/schemas/Company"},"rec":{"$ref":"#/components/schemas/Recognition"},"tagstats":{"type":"array","items":{"$ref":"#/components/schemas/TagStats"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["tagId","value","company","rec","tagstats"]},"Company":{"type":"object","properties":{"companyId":{"type":"number"},"name":{"type":"string"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}},"recognitions":{"type":"array","items":{"$ref":"#/components/schemas/Recognition"}},"users":{"type":"array","items":{"$ref":"#/components/schemas/Users"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["companyId","name","tags","recognitions","users"]}}},"paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"responses":{"201":{"description":""}}}},"/auth/refreshtoken":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"responses":{"201":{"description":""}}}},"/users/profile":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Users"}}}}}}},"/users/employeeId/{employ_id}":{"get":{"operationId":"UsersController_getUser","parameters":[{"name":"employ_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Users"}}}}}}},"/users/company/{comp_id}":{"get":{"operationId":"UsersController_getUsersByCompany","parameters":[{"name":"comp_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Users"}}}}}}}},"/users/{employeeId}":{"delete":{"operationId":"UsersController_removeUser","parameters":[{"name":"employeeId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/users/createDummy":{"post":{"operationId":"UsersController_createDummy","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Users"}}}}}}},"/users/create":{"post":{"operationId":"UsersController_createUser","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Users"}}}}}}},"/users/stats/{employeeId}":{"get":{"operationId":"UsersController_getStats","parameters":[{"name":"employeeId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/users/create_multiple":{"post":{"operationId":"UsersController_createUserMultiple","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/company":{"get":{"operationId":"CompanyController_getUserCompany","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Company"}}}}}}},"/company/all":{"get":{"operationId":"CompanyController_getCompany","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Company"}}}}}}}},"/company/{id}":{"get":{"operationId":"CompanyController_getOneCompany","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Company"}}}}}},"delete":{"operationId":"CompanyController_delete","parameters":[],"responses":{"200":{"description":""}}}},"/company/create":{"post":{"operationId":"CompanyController_createCompany","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Company"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Company"}}}}}}},"/rockstar":{"get":{"operationId":"RockstarController_getRockstar","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rockstar"}}}}}}},"/recognitions/all":{"get":{"operationId":"RecognitionController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Recognition"}}}}}}}},"/recognitions/create":{"post":{"operationId":"RecognitionController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recognition"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recognition"}}}}}}},"/recognitions/{id}":{"delete":{"operationId":"RecognitionController_delete","parameters":[],"responses":{"200":{"description":""}}}},"/recognitions/reports/{recID}":{"get":{"operationId":"RecognitionController_getReports","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/recognitions/report/{recID}":{"post":{"operationId":"RecognitionController_addReport","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Report"}}}}}}},"/recognitions/comments/{recID}":{"get":{"operationId":"RecognitionController_getComments","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/recognitions/comment/{recID}":{"post":{"operationId":"RecognitionController_addComment","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/recognitions/reactions/{recID}":{"get":{"operationId":"RecognitionController_getLikes","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/recognitions/reaction/{recID}":{"post":{"operationId":"RecognitionController_addReaction","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/recognitions/report/{reportID}":{"delete":{"operationId":"RecognitionController_deleteReport","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Report"}}}}}}}},"/recognitions/comment/{commentID}":{"delete":{"operationId":"RecognitionController_deleteComment","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Comment"}}}}}}}},"/recognitions/reaction/{reactionID}":{"delete":{"operationId":"RecognitionController_deleteReaction","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reaction"}}}}}}}},"/tag":{"get":{"operationId":"TagController_getCompanyTags","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}}}}}}}},"/tag/{id}":{"delete":{"operationId":"TagController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/tag/create":{"post":{"operationId":"TagController_create","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}}}}}}}